Use CLOCK_REALTIME for pthread_cond_timedwait
pthread_cond_timedwait default uses CLOCK_REALTIME internally for getting the
time. It is immune to time shifts - it just return prematurely.

pthread_condattr_setclock can be used to initialixe an attirbute for use with
pthread_mutex_init. It does not work so I just revert this specific clock
back to CLOCK_REALTIME.